[MERGE] Merge code with flexible column.

+++ src/Eccube/Controller/Admin/Product/CsvImportController.php
@@ -526,27 +526,30 @@ class CsvImportController
                         }
 
                         // カテゴリ削除フラグ対応
-                        if ($row['カテゴリ削除フラグ'] == '') {
-                            $Category->setDelFlg(Constant::DISABLED);
-                            $status = $app['eccube.repository.category']->save($Category);
-                        } else {
-                            if ($row['カテゴリ削除フラグ'] == (string)Constant::DISABLED || $row['カテゴリ削除フラグ'] == (string)Constant::ENABLED) {
-                                $Category->setDelFlg($row['カテゴリ削除フラグ']);
-                            } else {
-                                $this->addErrors(($data->key() + 1) . '行目のカテゴリ削除フラグが設定されていません。');
+                        if (isset($row['カテゴリ削除フラグ']) && Str::isNotBlank($row['カテゴリ削除フラグ'])) {
+                            $Category->setDelFlg($row['カテゴリ削除フラグ']);
+                            switch ($row['カテゴリ削除フラグ']) {
+                                case (string)Constant::DISABLED:
+                                    $status = $app['eccube.repository.category']->save($Category);
+                                    break;
+
+                                case (string)Constant::ENABLED:
+                                    $status = $app['eccube.repository.category']->delete($Category);
+                                    break;
+
+                                default:
+                                    $this->addErrors(($data->key() + 1) . '行目のカテゴリ削除フラグが設定されていません。');
+                                    return $this->render($app, $form, $headers, $this->categoryTwig);
+                                    break;
+                            }
+
+                            if (!$status) {
+                                $this->addErrors(($data->key() + 1) . '行目のカテゴリが、子カテゴリまたは商品が紐付いているため削除できません。');
                                 return $this->render($app, $form, $headers, $this->categoryTwig);
                             }
-
-                            if ($row['カテゴリ削除フラグ'] == (string)Constant::ENABLED) {
-                                $status = $app['eccube.repository.category']->delete($Category);
-
-                                if (!$status) {
-                                    $this->addErrors(($data->key() + 1) . '行目のカテゴリが、子カテゴリまたは商品が紐付いているため削除できません。');
-                                    return $this->render($app, $form, $headers, $this->categoryTwig);
-                                }
-                            } else {
-                                $status = $app['eccube.repository.category']->save($Category);
-                            }
+                        } else {
+                            $Category->setDelFlg(Constant::DISABLED);
+                            $status = $app['eccube.repository.category']->save($Category);
                         }
 
                         if (!$status) {
